SemesterBlock Commit
Fixed the CSS where the semester and its number would be next to each other instead I made it closer to the figma where its the semester on top then the number. Same thing for the type of semester where I put it next to the Semester number. And also moved the classes closer to the right, next time I just need to make it look better and closer to the figma.